making changes to habits and other class names- changed nav
