+++ TheHabitController/app/src/main/java/com/example/thehabitcontroller_project/EventsFragmentActivity.java
@@ -10,11 +10,12 @@ import android.view.ViewGroup;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link home#newInstance} factory method to
+ * Use the {@link EventsFragmentActivity#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class home extends Fragment {
+public class EventsFragmentActivity extends Fragment {
 
+    // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_PARAM1 = "param1";
     private static final String ARG_PARAM2 = "param2";
@@ -23,7 +24,7 @@ public class home extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public home() {
+    public EventsFragmentActivity() {
         // Required empty public constructor
     }
 
@@ -33,11 +34,11 @@ public class home extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment home.
+     * @return A new instance of fragment Events.
      */
     // TODO: Rename and change types and number of parameters
-    public static home newInstance(String param1, String param2) {
-        home fragment = new home();
+    public static EventsFragmentActivity newInstance(String param1, String param2) {
+        EventsFragmentActivity fragment = new EventsFragmentActivity();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -58,6 +59,6 @@ public class home extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_home, container, false);
+        return inflater.inflate(R.layout.fragment_events, container, false);
     }
 }
