+++ TheHabitController/app/src/main/java/com/example/thehabitcontroller_project/HomeFragmentActivity.java
@@ -10,12 +10,11 @@ import android.view.ViewGroup;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link Habits#newInstance} factory method to
+ * Use the {@link HomeFragmentActivity#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class Habits extends Fragment {
+public class HomeFragmentActivity extends Fragment {
 
-    // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_PARAM1 = "param1";
     private static final String ARG_PARAM2 = "param2";
@@ -24,7 +23,7 @@ public class Habits extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public Habits() {
+    public HomeFragmentActivity() {
         // Required empty public constructor
     }
 
@@ -34,11 +33,11 @@ public class Habits extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment Habits.
+     * @return A new instance of fragment home.
      */
     // TODO: Rename and change types and number of parameters
-    public static Habits newInstance(String param1, String param2) {
-        Habits fragment = new Habits();
+    public static HomeFragmentActivity newInstance(String param1, String param2) {
+        HomeFragmentActivity fragment = new HomeFragmentActivity();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -53,16 +52,12 @@ public class Habits extends Fragment {
             mParam1 = getArguments().getString(ARG_PARAM1);
             mParam2 = getArguments().getString(ARG_PARAM2);
         }
-        else {
-            mParam1 = "test";
-            mParam2 = "test2";
-        }
     }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_habits, container, false);
+        return inflater.inflate(R.layout.fragment_home, container, false);
     }
 }
