+++ TheHabitController/app/src/main/res/layout/fragment_events.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Events">
+    tools:context=".EventsFragmentActivity">
 
     <TextView
         android:layout_width="match_parent"
