+++ TheHabitController/app/src/main/res/layout/fragment_habits.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".Habits">
+    tools:context=".HabitsFragmentActivity">
 
     <ListView
         android:id="@+id/habit_list"
